testing: register pytest plugin
