+++ dvc/testing/benchmarks/cli/commands/test_plots.py
@@ -1,7 +1,9 @@
 import pytest
 
 from dvc.repo import Repo
-from tests.benchmarks.conftest import pull
+from dvc.testing.benchmarks.fixtures import _pull
+
+# pylint: disable=redefined-outer-name,unused-argument
 
 # pylint: disable=unused-argument
 
@@ -9,7 +11,7 @@ from tests.benchmarks.conftest import pull
 @pytest.mark.requires(minversion=(2, 34, 0), reason="top-level plots not supported")
 def test_plots(project, bench_dvc):
     with Repo() as dvc:
-        pull(dvc)
+        _pull(dvc)
 
     kwargs = {"rounds": 5, "iterations": 3, "warmup_rounds": 2}
     bench_dvc("plots", "show", name="show", **kwargs)
